Move sweep module and analyze script into chat_sl recipe
Sweep code and results now live together under recipes/chat_sl/:
- recipes/chat_sl/sweep/ — grid, run, collect, CLI, analyze script
- recipes/chat_sl/results/ — pre-computed SFT sweep tables + NLL plots

This avoids premature abstraction — when we have enough evidence from
RL/DPO sweeps, we can promote common parts to a shared module.
