+++ tinker_cookbook/recipes/chat_sl/sweep/__main__.py
@@ -1,0 +1,3 @@
+from tinker_cookbook.recipes.chat_sl.sweep.cli import main
+
+main()
