--- tinker_cookbook/sweep/__main__.py
@@ -1,3 +1,0 @@
-from tinker_cookbook.sweep.cli import main
-
-main()
